Upcast alignment probe to float32 (bf16 inputs broke raw measurement forward)
capture_initial/_capture_activations run the model without autocast, so a bf16
align_sample hit float32 weights -> dtype mismatch. Cast the 32-sample probe to
float32 (its pre-bf16 dtype); data buffers stay bf16.

+++ experiments/vision/train.py
@@ -305,7 +305,7 @@ def main() -> None:
             yb = yb.to(device, non_blocking=True)
 
             if dynamic and align_sample is None:
-                align_sample = xb[:args.sample_size].detach().clone().to(device)
+                align_sample = xb[:args.sample_size].detach().clone().float().to(device)
                 param.capture_initial(align_sample)
 
             optimizer.zero_grad()
